requirements.txt
mistralai>=0.4.0
faiss-cpu>=1.7.4
numpy>=1.24.0
python-dotenv>=1.0.0
docling>=1.0.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
streamlit>=1.28.0
langchain>=0.1.0
pypdf>=3.17.0
PyYAML>=6.0
SQLAlchemy>=2.0.0
python-docx>=0.8.11